+++ tests/test_core.py
@@ -1,9 +1,9 @@
 #
 # Copyright (C) 2015-2020 University of Oxford
 #
-# This file is part of utils.
+# This file is part of msprime.
 #
-# utils.is free software: you can redistribute it and/or modify
+# msprime is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
 # the Free Software Foundation, either version 3 of the License, or
 # (at your option) any later version.